+++ frontend/src/components/index.ts
@@ -11,9 +11,9 @@ export * from './forms/InquiryForm/InquiryForm';
 export * from './loading/Loading/Loading';
 export * from './stepper/StatusStepper';
 export * from './table/TableContainer/TableContainer';
-export * from './tablecells/DateCell';
-export * from './tablecells/RolesCell';
-export * from './tablecells/StatusCell';
-export * from './tabs/CustomTabPanel';
+export * from './tablecells/DataCell/DateCell';
+export * from './tablecells/RolesCell/RolesCell';
+export * from './tablecells/StatusCell/StatusCell';
+export * from './tabs/CustomeTabPanel/CustomTabPanel';
 export * from './table/InquiryTableContainer/InquiryTableContainer';
 export * from './table/UserTableContainer/UserTableContainer';
